+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 # Changelog: View
 
 > $Id$ ($Date$)
+
+## History
+
+### 0.1.1 @2022 Jul 29
+
+ - Added a shiny new PhpRenderer
+ - Added a shiny new StringRenderer
+ - Removed the deprecated classes
+ - Updated README
+
+### 0.1.0 @2022 Jul 26
+
+ - initial